Update order summary price fields to use Decimal type
* Change filter_factor parameter from float to Decimal in calculate_summaries and calculate_order_summary_detail
* Update OrderSummaries dataclass to use Decimal for filter_factor field
* Convert all price-related calculations to use Decimal("0.00") instead of 0.0
* Update five_percent_of_items calculation to use Decimal arithmetic
* Replace argus_response_models with ARM alias throughout for consistency
* Add Decimal import to test file and update test assertions to use Decimal values
* Ensure price field consistency across all order summary operations

+++ src/pfmsoft/eve_argus/models/esi/argus_response_models.py
@@ -210,7 +210,7 @@ class OrderSummaries(EsiModelBase):
     region_id: int
     solar_system_id: int | None
     location_id: int | None
-    filter_factor: float
+    filter_factor: Decimal
     summaries: dict[int, OrderSummary]
 
     def serialize(self, indent: int | None = 2) -> str:
